Read one glob dialect everywhere
Selectors were translated by fnmatch in filter_df and on string coords
and by a SQLite-shaped translator in the index, so a class negation was
spelled [!x] on one side and [^x] on the other and the same pattern
picked opposite halves depending on which side answered. dc.read of a
DASDAE file with a tag pattern took the fnmatch side while every spool
selection took the index side.

Move the SQLite translator to dascore.utils.misc, which both callers can
already import, and delete get_regex. The DOTALL flag moves into the
pattern text because pandas refuses a compiled pattern carrying flags it
did not set. Dropping the cycle also lets _spool_inventory import it at
module top rather than inside the function.
