+++ dascore/core/_spool_inventory.py
@@ -55,7 +55,7 @@ from dascore.exceptions import (
 from dascore.models import values_equal
 from dascore.units import convert_units, get_quantity_str
 from dascore.utils.intervals import interval_masks, value_kind
-from dascore.utils.misc import iterate, validate_acquisition_key
+from dascore.utils.misc import glob_to_regex, iterate, validate_acquisition_key
 from dascore.utils.time import to_datetime64
 
 # One vocabulary for both fiber verbs. What the quiet option leaves
@@ -537,8 +537,6 @@ def glob_filter(include, exclude):
     Globs mean what they mean everywhere else here, which is what SQLite
     means by them rather than what `fnmatch` does.
     """
-    from dascore.io.index.query import glob_to_regex  # noqa: PLC0415
-
     patterns = tuple(
         None if spec is None else [glob_to_regex(str(x)) for x in iterate(spec)]
         for spec in (include, exclude)
