+++ dascore/utils/misc.py
@@ -1509,3 +1509,79 @@ def validate_acquisition_key(value: str) -> str:
         )
         raise InvalidInventoryError(msg)
     return value
+
+
+@cache
+def glob_to_regex(pattern: str) -> re.Pattern:
+    """
+    Translate a glob to the regex which matches what SQLite's GLOB does.
+
+    SQLite is the authority on what a glob selector means, since that is
+    what the index applies, and it is not `fnmatch`: a character class is
+    negated with `[^...]`, where fnmatch spells that `[!...]` and reads a
+    leading `!` as a literal. Translating here rather than reaching for
+    fnmatch is what keeps one pattern from selecting opposite halves of a
+    spool depending on which side answered it. An unterminated class
+    matches nothing, as it does in SQLite; a class a regex cannot express
+    at all (a reversed range, which SQLite reads leniently) matches
+    nothing here rather than being guessed at.
+    """
+    out, index, size = [], 0, len(pattern)
+    while index < size:
+        char = pattern[index]
+        if char in "*?":
+            out.append(".*" if char == "*" else ".")
+        elif char == "[":
+            # A class may open with '^' to negate and may hold ']' as its
+            # first member; the class ends at the next ']' after those.
+            end = index + 1 + pattern[index + 1 : index + 2].count("^")
+            end += pattern[end : end + 1].count("]")
+            end = pattern.find("]", end)
+            if end < 0:
+                return _MATCHES_NOTHING
+            body = pattern[index + 1 : end]
+            negate, body = body.startswith("^"), body.removeprefix("^")
+            # A ']' opening a class is one of its members, and SQLite takes
+            # it as a plain one: it is not the low end of a range, so the
+            # dash which may follow it is a member too.
+            leading = ""
+            if body.startswith("]"):
+                leading, body = re.escape("]"), body[1:]
+            out.append(f"[{'^' if negate else ''}{leading}{_class_body(body)}]")
+            index = end
+        else:
+            out.append(re.escape(char))
+        index += 1
+    # The DOTALL flag is written into the pattern, since SQLite's wildcards
+    # cross a newline like any other byte and pandas rejects a compiled
+    # pattern whose flags it did not set itself.
+    return re.compile(r"(?s)" + "".join(out) + r"\Z")
+
+
+# A pattern which matches nothing, for a glob SQLite would not read.
+_MATCHES_NOTHING = re.compile(r"(?!)")
+
+
+def _class_body(body: str) -> str:
+    """
+    Escape the members of a glob character class for a regex one.
+
+    Ranges stay ranges, since a glob class means them, with each endpoint
+    escaped on its own — escaping the body wholesale would turn the dash
+    of a range into a member. A range's low endpoint is also emitted as a
+    member: SQLite tests it before testing the range, so `[z-a]` matches
+    `z` where the reversed range alone matches nothing.
+    """
+    out, index, size = [], 0, len(body)
+    while index < size:
+        low = body[index]
+        if index + 2 < size and body[index + 1] == "-":
+            high = body[index + 2]
+            out.append(re.escape(low))
+            if low <= high:
+                out.append(f"{re.escape(low)}-{re.escape(high)}")
+            index += 3
+            continue
+        out.append(re.escape(low))
+        index += 1
+    return "".join(out)
